requirements.txt
python-dotenv
flask-sqlalchemy
flask
flask-socketio
requests